Chore: 2차 재모듈화
