+++ Service/model/Embedding_Chatbot.py
@@ -1,15 +1,18 @@
 from openai import OpenAI
 # openai API
 
-from ..config.Authentication import Authentication
-from data.redis_data.RedisManager import RedisVectorStore
+from config.Authentication import Authentication
 # Modules
 
+from redis import Redis
+from redis.commands.search.query import Query
+import numpy as np
 from typing import List
 
+# packages
 
 class Embedding_Chatbot:
-    def __init__(self):
+    def __init__(self, index_name='artist_vector_store'):
         self.client = OpenAI(
             api_key=Authentication().get_token()
         )
@@ -17,7 +20,10 @@ class Embedding_Chatbot:
         self.embedding_model = self.client\
             .embeddings\
             .create
-        self.vector_store = RedisVectorStore()
+
+        # VectorStore
+        self.redis_conn = Redis()
+        self.index_name = index_name
 
         # 기초 model parts 생성
     def get_embedding(self, text: str) -> List[float]:
@@ -27,9 +33,42 @@ class Embedding_Chatbot:
             model = model
         ).data[0].embedding
 
+    def search_similar_artist(self, query_text: str, top_k: int = 3):
+        """
+        입력 테스트와 유사한 아티스트 벡터를 검색
+        :param query_text:
+        :param top_k:
+        :return:
+        """
+
+        query_response = self.client.embeddings.create(
+            input=[query_text],
+            model='text-embedding-3-small'
+        )
+        query_embedding = query_response.data[0].embedding
+
+        np_query_embedding = np.array(query_embedding, dtype=np.float32).tobytes()
+
+        query = f"*=>[KNN {top_k} @embedding $vec_param AS dist]"
+        params_dict = {
+            "vec_param" : np_query_embedding,
+        }
+        try:
+            # Query 객체 생성 및 다이얼렉트 설정
+            q = Query(query) \
+                .return_fields("content","이름", "그룹", "특징", "dist") \
+                .sort_by("dist") \
+                .paging(0, top_k) \
+                .dialect(2)  # 다이얼렉트 2로 설정
+
+            results = self.redis_conn.ft(self.index_name).search(q, query_params=params_dict)
+            return results.docs  # 검색된 문서 리스트 반환
+        except Exception as e:
+            print(f"Error during search: {e}")
+            return []
+
     def create_context(self, question: str, max_len: int = 1800) -> str:
-        simliar_docs = self.vector_store\
-            .search_similar_artist(question, top_k= 5)
+        simliar_docs = self.search_similar_artist(question, top_k= 5)
         if not simliar_docs:
             return ""
         cur_len = 0
@@ -43,12 +82,11 @@ class Embedding_Chatbot:
             cur_len += additional_length
         return "\n".join(context_parts)
 
+
     def answer_question(self, question: str, max_len: int = 1800) -> str:
         context = self.create_context(question, max_len=max_len)
         system_message = """
-        당신은 음악 전문가이자 작사가입니다. 사용자 질문에 따라 적절한 답변을 제공해야 합니다.
-        - 음악 및 특정 데이터에 대한 정보 조회 요청 시 데이터베이스의 정보를 기반으로 질문에 대한 정확한 답변만 제공합니다(제공된 데이터 종류: 이름, 본명, 생년월일, 데뷔, 그룹, 히트곡, 특징)        
-        - 노래 가사 생성 요청 시, 위에서 제공해준 데이터를 기반으로 창의적인 가사를 생성해줍니다.(무조건 생성이라는 질문에만 답변을 줘야함)        
+        당신은 음악 전문가이자 작사가입니다. 사용자 질문에 따라 적절한 답변을 제공해야 합니다. 
         """
 
         if context:
@@ -79,3 +117,8 @@ class Embedding_Chatbot:
         except Exception as e:
             print(f"응답 생성 중 오류 발생: {e}")
             return "죄송합니다. 답변을 생성할 수 없습니다."
+
+
+
+
+
